Rename Tapfiliate API client to client and update imports.
