+++ apps/web/lib/tapfiliate/import-customers.ts
@@ -7,7 +7,7 @@ import { syncPartnerLinksStats } from "../api/partners/sync-partner-links-stats"
 import { recordClick, recordLeadWithTimestamp } from "../tinybird";
 import { logImportError } from "../tinybird/log-import-error";
 import { clickEventSchemaTB } from "../zod/schemas/clicks";
-import { TapfiliateClient } from "./api";
+import { TapfiliateClient } from "./client";
 import { TAPFILIATE_MAX_BATCHES, tapfiliateImporter } from "./importer";
 import { TapfiliateCustomer, TapfiliateImportPayload } from "./types";
